requirements.txt
streamlit
PyMuPDF
spire.pdf
camelot-py[cv]
matplotlib
numpy
reportlab
Pillow
transformers
torch
requests
google-generativeai 